[Refaactor] ProjectDescriptionHelpers 폴더링
