+++ Livith-iOS/Tuist/ProjectDescriptionHelpers/Module/Module+TargetID.swift
@@ -1,27 +1,11 @@
 //
-//  Module.swift
+//  TargetID.swift
 //  Manifests
 //
-//  Created by Youjin Lee on 10/11/25.
+//  Created by Youjin Lee on 11/2/25.
 //
 
 import ProjectDescription
-
-// MARK: - ProjectID
-
-public enum ProjectID: String, CaseIterable {
-    case app = "App"
-    case core = "Core"
-    case shared = "Shared"
-    case search = "Search"
-    case login = "Login"
-    
-    public var name: String { rawValue }
-    
-    public var path: Path {
-        .relativeToRoot("Projects/\(rawValue)")
-    }
-}
 
 // MARK: - TargetID
